+++ src/helm/proxy/services/server_service.py
@@ -49,7 +49,7 @@ class ServerService(Service):
 
         self.client = AutoClient(credentials, cache_path, mongo_uri)
         self.tokenizer = AutoTokenizer(credentials, cache_path, mongo_uri)
-        self.token_counter = AutoTokenCounter(self.client.get_huggingface_tokenizer())
+        self.token_counter = AutoTokenCounter(self.tokenizer.get_huggingface_tokenizer())
         self.accounts = Accounts(accounts_path, root_mode=root_mode)
         # Lazily instantiated by get_toxicity_scores()
         self.toxicity_classifier_client: Optional[ToxicityClassifierClient] = None
