+++ src/helm/proxy/token_counters/openai_token_counter.py
@@ -15,7 +15,6 @@ class OpenAITokenCounter(TokenCounter):
         Counts the total number of tokens using the suggestion here:
         https://community.openai.com/t/how-do-i-calculate-the-pricing-for-generation-of-text/11662/5
         """
-
         tokenized_prompt: TokenizationRequestResult = self.huggingface_tokenizer.tokenize(
             TokenizationRequest(request.prompt)
         )
